Update utils.py
So, as I've noticed, the Steam balance is currently displayed like this: '1 234.5$'. So, this results that the user is receiving a balance value in the format '1' - (the first numbers of his balance, before the space). This can be fixed by simply adding space removal to the "parse_price" function. Perhaps, it'd be better if anyone find alternative, simpler, and more universal solution for such situations.